requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==3.7.1
assemblyai==0.30.0
astunparse==1.6.3
audioread==3.0.1
cachetools==5.3.3
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==7.1.2
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
decorator==4.4.2
exceptiongroup==1.2.2
ffmpeg-python==0.2.0
filelock==3.15.4
flatbuffers==1.12
fonttools==4.53.1
fsspec==2024.6.1
future==1.0.0
gast==0.4.0
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==0.4.6
google-generativeai==0.7.2
google-pasta==0.2.0
googleapis-common-protos==1.63.2
grpcio==1.64.1
grpcio-status==1.62.2
h11==0.12.0
h2==4.1.0
h5py==3.11.0
hpack==4.0.0
httpcore==0.13.7
httplib2==0.22.0
httpx==0.19.0
hyperframe==6.0.1
idna==3.7
imageio==2.34.2
imageio-ffmpeg==0.5.1
intel-openmp==2021.4.0
Jinja2==3.1.4
joblib==1.4.2
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2.post1
llvmlite==0.43.0
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.8.4
mkl==2021.4.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.8
networkx==3.3
norbert==0.2.1
numba==0.60.0
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.8.1.78
opt-einsum==3.3.0
packaging==24.1
pandas==1.5.3
pillow==10.4.0
platformdirs==4.2.2
pooch==1.8.2
proglog==0.1.10
proto-plus==1.24.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytube==15.0.0
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
resampy==0.4.3
rfc3986==1.5.0
rsa==4.9
scikit-learn==1.5.1
scipy==1.14.0
six==1.16.0
sniffio==1.3.1
soundfile==0.12.1
soxr==0.3.7
spleeter==2.4.0
sympy==1.13.0
tbb==2021.13.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.3
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.5.0
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
tqdm==4.66.4
typer==0.3.2
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
websockets==12.0
Werkzeug==3.0.3
wrapt==1.16.0
